+++ R/datatables.R
@@ -270,6 +270,8 @@ datatable = function(
     if (identical(options$lengthMenu, c(10, 25, 50, 100)))
       options$lengthMenu = NULL  # that is just the default
   }
+  if (!is.null(options[['search']]) && !is.list(options[['search']]))
+    stop("The value of `search` in `options` must be NULL or a list")
 
   # record fillContainer and autoHideNavigation
   if (!is.null(fillContainer)) params$fillContainer = fillContainer