+++ R/datatables.R
@@ -290,6 +290,8 @@ datatable = function(
     if (identical(options$lengthMenu, c(10, 25, 50, 100)))
       options$lengthMenu = NULL  # that is just the default
   }
+  if (!is.null(options[['search']]) && !is.list(options[['search']]))
+    stop("The value of `search` in `options` must be NULL or a list")
 
   # record fillContainer and autoHideNavigation
   if (!is.null(fillContainer)) params$fillContainer = fillContainer
@@ -768,9 +770,11 @@ DT2BSClass = function(class) {
     'cell-border' = 'table-bordered', 'compact' = 'table-condensed',
     'hover' = 'table-hover', 'stripe' = 'table-striped'
   )
+  # translate known default styling classes to BS table classes and keep
+  # unknown classes as they are
   class = c(
     BSclass[intersect(class, names(BSclass))],
-    grep('^table-', class, value = TRUE)
+    setdiff(class, names(BSclass))
   )
   class = unique(c('table', class))
   paste(class, collapse = ' ')